update android project files
This should fix all the kotlin and agp version incompatibilities peoples
were having.

This updates the AGP version to the latest version which still works
with java 8. Bump gradle version to the minimum required.

Update kotlin to latest 1.4 point release release.

Replace jcenter with mavenCentral as jcenter is scheduled for shutdown.