refactor!: remove unused playerStateStream (#1280)
Remove unused player state stream. See this comment.

BREAKING CHANGE: But it should not play a role:

stream was never called, so it never emitted any events, (except for web, which also already emits onComplete event). So most likely no-one did listen to this stream.
the according stream is further exposed in main audioplayers package, so it's not breaking the main package.