fix(android): reset prepared state on player error (#1425)
In the rare case, that an media player error occurs (like when playing
unsupported media sources), the player still is in the "prepared" state,
although the internal MediaPlayer state isn't anymore. Then it could keep
calling methods, which lead to an illegal MediaPlayer state. By
resetting this variable on a media error, this should not make the
library crash anymore, but still emit the error.

Fixes #1260
Fixes #1331