docs: Fix LICENSE files for windows and linux (#1431)
Flipping through pub I figured that unlike the AP-base package and all
other AP packages, the Windows and Linux packages did not seem to have a
LICENSE file associated.


![image](https://user-images.githubusercontent.com/882703/224607336-cbc6432f-d5b2-4967-9644-dff3a33c49d3.png)

Looking at GitHub, it was a simple mistake - the "symlink" was copied in
a way that didn't propagate the correct file type when those packages
were created.

They actually had physical files with the path of the symlink written on
them. A fresh `ln -s` quickly amended the issue. The diff on GitHub web
interface might look weird, but if you check git, what really changed
was just the file type.

This should fix the pub issues with those two packages on the next
release.

Before:

![image](https://user-images.githubusercontent.com/882703/224607277-9214d6c1-95dd-4446-999a-021eebe93728.png)

Now:

![image](https://user-images.githubusercontent.com/882703/224607253-ca4f9061-241e-4915-912b-3e46d5725584.png)