ci: Use fixed version of flutter-action (#1506)
# Description

Like in
[flame](https://github.com/flame-engine/flame/blob/19d1e72e4bb51550752b346525e51be67edc2019/.github/workflows/cicd.yml#L18)
it is more practicable to use fixed version of `flutter-action` to avoid
unexpected failures of the CI.

Further when flutter is releasing a new version, one should upgrade the
version of `flutter-action` and also add the required changes to pass
the tests at the same time.

This ensures that the tests in the main always stay consistent. It also
makes the development faster and more independent of other packages such
as `flame-lint`.