test: Attempt to fix flaky tests for Windows & Android (#1441)
# Description

The tests in Windows & Android are possibly flaky for this reason:
- In general it is not tested, if a source starts to play for app tests.
- In the app test for `ReleaseMode.loop`, the source starts to play, but
never completes or stops. See:
https://github.com/bluefireteam/audioplayers/blob/fe81f3b1e600fe005febbe7cd3da02735a3de004/packages/audioplayers/example/integration_test/app/tabs/controls_tab.dart#L104
. This is now checked by getting the current position.

The solution to the problem also raises another issue on Android, as the
position can not be retrieved on very short sources, thus they are
excluded from these tests and it is expected to make it work with
Exoplayer (#1526).