ci: Set system date time after downloading virtual audio device (#1770)
# Description

On executing the windows tests, a virtual audio device is needed, which
can only be installed with a certificate valid to a specific date.
Recently the tests failed, as the date time was set before downloading
and therefore threw a `NotTimeValid` error. This fixes that by
downloading the source first, setting the date time and then installing
it.